requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.0
python-dotenv==1.0.1
pytest==8.3.0
openai==1.58.1
